Refactored App Name
