+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.locationactivity"
+        applicationId "com.example.potholedetectionapp"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
