+++ app/src/androidTest/java/com/example/potholedetectionapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity;
+package com.example.potholedetectionapp;
 
 import android.content.Context;
