+++ app/src/main/java/com/example/potholedetectionapp/DatabaseHelper.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity;
+package com.example.potholedetectionapp;
 
 import android.content.ContentValues;
 import android.content.Context;
