+++ app/src/main/java/com/example/potholedetectionapp/Datapoint.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity;
+package com.example.potholedetectionapp;
 
 import android.os.Parcel;
 import android.os.Parcelable;
