+++ app/src/main/java/com/example/potholedetectionapp/libsvm/svm.java
@@ -3,7 +3,7 @@
 
 
 
-package com.example.locationactivity.libsvm;
+package com.example.potholedetectionapp.libsvm;
 import java.io.*;
 import java.util.*;
