+++ app/src/main/java/com/example/potholedetectionapp/libsvm/svm_node.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity.libsvm;
+package com.example.potholedetectionapp.libsvm;
 public class svm_node implements java.io.Serializable
 {
 	public int index;
