+++ app/src/main/java/com/example/potholedetectionapp/libsvm/svm_print_interface.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity.libsvm;
+package com.example.potholedetectionapp.libsvm;
 public interface svm_print_interface
 {
 	public void print(String s);
