+++ app/src/main/java/com/example/potholedetectionapp/libsvm/svm_problem.java
@@ -1,4 +1,4 @@
-package com.example.locationactivity.libsvm;
+package com.example.potholedetectionapp.libsvm;
 public class svm_problem implements java.io.Serializable
 {
 	public int l;
