+++ app/src/main/java/com/example/potholedetectionapp/svm_toy.java
