+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "LocationActivity"
+rootProject.name = "pothole-detection-app"
